refactor(whispering): fold shortcut sync/reset into operations/shortcuts
register-commands.ts registered no commands: it held the local/global
shortcut sync and reset-on-duplicate operations, built on the localShortcuts
primitive that already lives in $lib/operations/shortcuts. The name described
nothing it did, and the module was route-private under _runtime/ even though
two settings pages call it as live operations.

Fold it into $lib/operations/shortcuts next to localShortcuts so the whole
shortcut-operations surface has one home, and drop the now-internal cross
import. Renamed the duplicate-tracking Map in the local reset helper from
localShortcuts to seen, matching its global sibling and no longer shadowing
the localShortcuts primitive now in the same file.
